@@ -67,6 +67,8 @@
     "build:ts": "react-scripts-ts build",
     "deploy": "npm run build && gh-pages -d build",
     "test": "npm run build:tekster && react-scripts-ts test --env=jsdom",
+    "test:integration": "node integration/runner.js --debug-mode",
+    "test:browsers": "node integration/browserstack-runner.js",
     "lint": "tslint ./src/**/*.ts*",
     "lint:fix": "tslint ./src/**/*.ts* --fix",
     "test:coverage": "npm test -- --coverage",
@@ -76,20 +78,18 @@
     "start:tekster": "npm run build:tekster",
     "start": "npm-run-all -p start:*",
     "build": "npm run delete-static && npm run build:tekster && npm run build:ts",
-    "integrationtest": "cross-env REACT_APP_MOCK_BES=true node integration/nightwatch.js -c integration/default.conf.js",
     "integrationtest-build": "node scripts/preHeroku.js && npm install && cross-env REACT_APP_MOCK_BES=true cross-env REACT_APP_MOCK='true' npm run build && node scripts/postHeroku.js",
-    "test:browserstack": "node integration/browserstack.runner.js -c integration/browserstack.conf.js -e win_chrome,android_chrome,win_ie,win_edge",
     "heroku-serve": "node index.js",
     "heroku-postbuild": "node scripts/preHeroku.js && npm install && cross-env REACT_APP_MOCK_ENDRE_REG_LOP=true cross-env REACT_APP_MOCK='true' npm run build && node scripts/postHeroku.js"
   },
   "devDependencies": {
-    "@navikt/create-react-app-typescript-less": "^1.0.0-beta.14",
-    "@types/query-string": "5.1.0",
+    "@navikt/create-react-app-typescript-less": "^1.0.0-beta.16",
     "@types/classnames": "2.2.3",
     "@types/fetch-mock": "5.12.2",
     "@types/jest": "22.2.3",
     "@types/lodash": "4.14.108",
     "@types/node": "10.0.3",
+    "@types/query-string": "5.1.0",
     "@types/react": "16.3.18",
     "@types/react-dom": "16.0.5",
     "@types/react-intl": "2.3.8",
@@ -97,24 +97,24 @@
     "@types/react-router-dom": "4.2.6",
     "@types/react-select": "^1.2.0",
     "@types/react-test-renderer": "16.0.1",
+    "babel-core": "6.24.1",
     "chai": "4.1.2",
     "chai-as-promised": "7.1.1",
     "chai-spies": "^1.0.0",
     "chromedriver": "^2.42.0",
+    "cross-env": "^5.2.0",
     "enzyme": "3.3.0",
     "enzyme-adapter-react-16": "1.1.1",
     "enzyme-react-intl": "2.0.0",
+    "express": "^4.16.3",
     "fetch-mock": "5.13.1",
     "less": "3.0.2",
     "npm-run-all": "4.1.2",
     "path-to-regexp": "2.2.1",
     "sinon": "5.0.3",
-    "typescript": "^2.8.3",
-    "nightwatch": "^0.9.21",
+    "testcafe": "^0.23.0",
+    "testcafe-browser-provider-browserstack": "^1.5.0",
     "tree-kill": "^1.2.0",
-    "cross-env": "^5.2.0",
-    "express": "^4.16.3",
-    "browserstack-local": "^1.3.4",
-    "babel-core": "6.24.1"
+    "typescript": "^3.0.3"
   }
 }
