@@ -10,7 +10,6 @@ export enum ActionTypes {
 
 export interface Data {
     'arbeidssokerregistrering.gradual-rollout-ny-registrering': boolean;
-    'arbeidssokerregistrering.sykefravaer': boolean;
 }
 
 export interface State {
@@ -24,13 +23,12 @@ interface Action {
 }
 
 export const alleFeatureToggles = [
-    'arbeidssokerregistrering.gradual-rollout-ny-registrering', 'arbeidssokerregistrering.sykefravaer'
+    'arbeidssokerregistrering.gradual-rollout-ny-registrering'
 ];
 
 const initialState = {
     data : {
         'arbeidssokerregistrering.gradual-rollout-ny-registrering': false,
-        'arbeidssokerregistrering.sykefravaer': false
     },
     status: STATUS.NOT_STARTED
 };
@@ -66,8 +64,4 @@ export function selectFeatureToggles(state: AppState): Data {
 
 export function selectGradualRolloutNyRegistreringFeatureToggle(state: AppState): boolean {
     return state.featureToggles.data['arbeidssokerregistrering.gradual-rollout-ny-registrering'];
-}
-
-export function selectSykefravaerFeatureToggle(state: AppState): boolean {
-    return state.featureToggles.data['arbeidssokerregistrering.sykefravaer'];
 }
