@@ -7,7 +7,7 @@ import oversettelseAvStillingFraAAReg from './oversettelse-av-stilling-fra-aareg
 import {featureTogglesMock} from "./feature-toggles-mock";
 import {FEATURE_URL, VEILARBREGISTRERING_URL} from '../ducks/api';
 import autentisert from './autentiseringsinfo-mock';
-import registreringRespons from "./registrerbruker-mock";
+import { ordinaerRegistreringRespons, sykmeldtRegistreringRespons } from "./registrerbruker-mock";
 import getStore from "../store";
 import {ActionTypes as SvarActionTypes, SporsmalId} from '../ducks/svar';
 import svarMock from "./svar-mock";
@@ -80,8 +80,11 @@ if(MOCK_STYRK08_PAMJANZZ) {
 }
 
 if (MOCK_REGISTRER_BRUKER) {
-    const response = respondWith(delayed(DELAY, registreringRespons, 200));
-    (mock as any).post(`${VEILARBREGISTRERING_URL}/startregistrering`, response);
+    const ordinaerRespons = respondWith(delayed(DELAY, ordinaerRegistreringRespons, 200));
+    const sykmeldtRespons = respondWith(delayed(DELAY, sykmeldtRegistreringRespons, 200));
+
+    (mock as any).post(`${VEILARBREGISTRERING_URL}/startregistrering`, ordinaerRespons);
+    (mock as any).post(`${VEILARBREGISTRERING_URL}/startregistrersykmeldt`, sykmeldtRespons);
 }
 
 if (MOCK_REAKTIVER_BRUKER) {
